requirements.txt
streamlit>=1.41.0,<2.0
snowflake-connector-python>=3.6.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=1.5.0
altair>=5.0.0,<6.0
toml>=0.10.0
